Implement the needed functionality for Fleek Network
We also remove the platform specific implementation of the compress
function and depend on the actual `blake3` crate to provide that.

This prevents linking errors that might come up because of multiple
definition of those native codes.

Additionally, since we depend on the actual `blake3` crate we forward
all of the cargo features to the upstream dependency and remove the
`build.rs` file which is not needed anymore.